+++ frontend/messages/zh.js
@@ -5,6 +5,22 @@ export default [
     "defaultMessage": "刷新",
   },
   {
+    "id": "app.errors.badRequest",
+    "defaultMessage": "The request parameters are incorrect, please check your input.",
+  },
+  {
+    "id": "app.errors.invalidData",
+    "defaultMessage": "The data is invalid, please try again.",
+  },
+  {
+    "id": "app.errors.network",
+    "defaultMessage": "Connection timeout, please network connection.",
+  },
+  {
+    "id": "app.errors.notFound",
+    "defaultMessage": "The requested resource not exists",
+  },
+  {
     "id": "app.errors.server",
     "defaultMessage": "Sorry, we've had a server error, Please try again.",
   },